[Bronze V] Title: A/B, Time: 208 ms, Memory: 17808 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.java
@@ -8,8 +8,7 @@ public class Main {
 		int b = scan.nextInt();
 		
 		if (a > 0 && b < 10) {
-			System.out.println((double)a / b);
+			System.out.println((double)a / (double)b);
 		}
-	}
-
+    }
 }
